requirements.txt
pillow
fpdf2
pyinstaller